refactor(server): group admin-console resources under http/console + usecases/console
Moves the wholly-admin resources into a console/ subdirectory in both http/ and
usecases/: storages, users, email-config, audit (handler + its dedicated
usecase), plus the teams-admin and licensing-admin handlers. The latter two keep
their usecases in usecases/ because team.ts and licensing.ts are shared with
user-facing routes.

Mixed resources that expose both admin and public/user endpoints (quotas,
branding, auth-providers, announcements, invite-codes, site-invitations,
downloaders) are intentionally left in place — splitting them would re-fragment
the one-usecase-per-resource consolidation.

Pure file moves + import-path updates; no behavior change. lint:http still
"fully locked", lint:arch clean, biome clean, full suite green (4327 passed).
