+++ server/app.ts
@@ -5,21 +5,24 @@ import { cors } from 'hono/cors'
 import type { Auth } from './auth'
 import { createDeps } from './composition'
 import { adminAnnouncements, announcements } from './http/announcements'
-import { adminAudit } from './http/audit'
 import { adminAuthProviders, publicAuthProviders } from './http/auth-providers'
 import backgroundJobs from './http/background-jobs'
 import { brandingAdmin, publicBranding } from './http/branding'
 import { cloudStore, cloudStoreWebhooks } from './http/cloud-store'
+import { adminAudit } from './http/console/audit'
+import emailConfig from './http/console/email-config'
+import licensingAdmin from './http/console/licensing-admin'
+import storages from './http/console/storages'
+import { adminTeams } from './http/console/teams-admin'
+import users from './http/console/users'
 import downloadTasks from './http/download-tasks'
 import downloaders, { downloaderSelfRoute } from './http/downloaders'
-import emailConfig from './http/email-config'
 import { events } from './http/events'
 import ihost from './http/ihost'
 import ihostConfig from './http/ihost-config'
 import internal from './http/internal'
 import { adminInviteCodes, publicInviteCodes } from './http/invite-codes'
 import licensing from './http/licensing'
-import licensingAdmin from './http/licensing-admin'
 import { me } from './http/me'
 import { notifications } from './http/notifications'
 import objects from './http/objects'
@@ -28,12 +31,9 @@ import { adminQuotas, userQuotas } from './http/quotas'
 import redirect from './http/redirect'
 import { authedShares, publicShares } from './http/shares'
 import { adminSiteInvitations, publicSiteInvitations } from './http/site-invitations'
-import storages from './http/storages'
 import system from './http/system'
 import { publicTeams, teams } from './http/teams'
-import { adminTeams } from './http/teams-admin'
 import trash from './http/trash'
-import users from './http/users'
 import webdav from './http/webdav'
 import { formatError } from './lib/errors'
 import { authMiddleware } from './middleware/auth'
