+++ server/http/console/users.ts
@@ -1,8 +1,8 @@
 import { zValidator } from '@hono/zod-validator'
 import { Hono } from 'hono'
 import { z } from 'zod'
-import { requireAdmin } from '../middleware/auth'
-import type { Env } from '../middleware/platform'
+import { requireAdmin } from '../../middleware/auth'
+import type { Env } from '../../middleware/platform'
 import {
   deleteUser,
   deleteUsers,
@@ -14,7 +14,7 @@ import {
   setUserStatus,
   setUsersStatus,
   updateUserEntitlement,
-} from '../usecases/user'
+} from '../../usecases/console/user'
 
 const updateStatusSchema = z.object({
   status: z.enum(['active', 'disabled']),
