+++ server/usecases/console/audit.ts
@@ -2,7 +2,7 @@
 // activity events. A single-port operation today; it lives here so the resource
 // has one home and the http handler stays free of deps access.
 
-import type { ActivityRepo, AdminAuditEventWithOrg, ListAdminAuditOpts } from './ports'
+import type { ActivityRepo, AdminAuditEventWithOrg, ListAdminAuditOpts } from '../ports'
 
 export function listAuditEvents(
   deps: { activity: Pick<ActivityRepo, 'listAdminAudit'> },
